Eliminating few signed/unsigned comparison warnings
git-svn-id: https://clucene.svn.sourceforge.net/svnroot/clucene/branches/lucene2_3_2@2784 20ef185c-fe11-0410-a618-ba9304b01011

+++ src/core/CLucene/search/Hits.cpp
@@ -149,7 +149,7 @@ CL_NS_DEF(search)
 				// either we cannot count deletions, or some "previously valid hits" might have been deleted, so find exact start point
 				nDeletedHits = 0;
 				debugCheckedForDeletions = true;
-				int32_t i2 = 0;
+				size_t i2 = 0;
 				for (size_t i1=0; i1<hitDocs->size() && i2 < scoreDocsLength; i1++) {
 					int32_t id1 = ((*hitDocs)[i1])->id;
 					int32_t id2 = scoreDocs[i2].doc;
